rest-python/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
Authlib==0.11
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
defusedxml==0.6.0
Flask==1.0.3
Flask-RESTful==0.3.7
httplib2==0.13.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==2.1.0
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
six==1.12.0
urllib3==1.25.3
URLObject==2.4.3
Werkzeug==0.15.4
zeep==3.4.0
